repos/walleria-simoes/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
yfinance>=0.2.37
pandas>=2.2.0
python-dotenv>=1.0.1
google-genai>=0.1.0
openpyxl>=3.1.2
streamlit>=1.31.0